Extract keybindings related types to `vs/base/common/keybindings`
