+++ src/vs/base/browser/ui/dropdown/dropdownActionViewItem.ts
@@ -13,7 +13,8 @@ import { DropdownMenu, IActionProvider, IDropdownMenuOptions, ILabelRenderer } f
 import { Action, IAction, IActionRunner } from 'vs/base/common/actions';
 import { Codicon } from 'vs/base/common/codicons';
 import { Emitter } from 'vs/base/common/event';
-import { KeyCode, ResolvedKeybinding } from 'vs/base/common/keyCodes';
+import { KeyCode } from 'vs/base/common/keyCodes';
+import { ResolvedKeybinding } from 'vs/base/common/keybindings';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import 'vs/css!./dropdown';
