+++ src/vs/base/common/scanCode.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IBaseKeybinding, KeyCode } from 'vs/base/common/keyCodes';
+import { KeyCode } from 'vs/base/common/keyCodes';
+import { IBaseKeybinding } from 'vs/base/common/keybindings';
 
 /**
  * keyboardEvent.code
