+++ src/vs/editor/common/standalone/standaloneEnums.ts
@@ -377,7 +377,6 @@ export enum InlineCompletionTriggerKind {
 	 */
 	Explicit = 1
 }
-
 /**
  * Virtual Key Codes, the value does not hold any inherent meaning.
  * Inspired somewhat from https://msdn.microsoft.com/en-us/library/windows/desktop/dd375731(v=vs.85).aspx
