+++ src/vs/platform/driver/electron-main/driver.ts
@@ -6,7 +6,8 @@
 import { timeout } from 'vs/base/common/async';
 import { Emitter, Event } from 'vs/base/common/event';
 import { KeybindingParser } from 'vs/base/common/keybindingParser';
-import { KeyCode, SimpleKeybinding } from 'vs/base/common/keyCodes';
+import { KeyCode } from 'vs/base/common/keyCodes';
+import { SimpleKeybinding } from 'vs/base/common/keybindings';
 import { combinedDisposable, IDisposable } from 'vs/base/common/lifecycle';
 import { OS } from 'vs/base/common/platform';
 import { ScanCodeBinding } from 'vs/base/common/scanCode';
