+++ src/vs/workbench/services/keybinding/test/electron-browser/keyboardMapperTestUtils.ts
@@ -6,7 +6,7 @@
 import * as assert from 'assert';
 import * as path from 'vs/base/common/path';
 import { getPathFromAmdModule } from 'vs/base/test/node/testUtils';
-import { Keybinding, KeybindingModifier, ResolvedKeybinding, SimpleKeybinding } from 'vs/base/common/keyCodes';
+import { Keybinding, KeybindingModifier, ResolvedKeybinding, SimpleKeybinding } from 'vs/base/common/keybindings';
 import { ScanCodeBinding } from 'vs/base/common/scanCode';
 import { Promises } from 'vs/base/node/pfs';
 import { IKeyboardEvent } from 'vs/platform/keybinding/common/keybinding';
